+++ moVirt/src/main/java/org/ovirt/mobile/movirt/ui/hosts/HostsFragment.java
@@ -10,7 +10,7 @@ import android.widget.TextView;
 import org.androidannotations.annotations.EFragment;
 import org.ovirt.mobile.movirt.R;
 import org.ovirt.mobile.movirt.model.Host;
-import org.ovirt.mobile.movirt.ui.ClusterBoundBaseEntityListFragment;
+import org.ovirt.mobile.movirt.ui.listfragment.ClusterBoundBaseEntityListFragment;
 
 import static org.ovirt.mobile.movirt.provider.OVirtContract.Host.CPU_USAGE;
 import static org.ovirt.mobile.movirt.provider.OVirtContract.Host.MEMORY_USAGE;
