+++ moVirt/src/main/java/org/ovirt/mobile/movirt/ui/vms/VmNicsFragment.java
@@ -16,8 +16,7 @@ import org.ovirt.mobile.movirt.R;
 import org.ovirt.mobile.movirt.auth.properties.manager.AccountPropertiesManager;
 import org.ovirt.mobile.movirt.model.Nic;
 import org.ovirt.mobile.movirt.ui.ProgressBarResponse;
-import org.ovirt.mobile.movirt.ui.SnapshotEmbeddableVmBoundResumeSyncableBaseEntityListFragment;
-import org.ovirt.mobile.movirt.ui.VmBoundResumeSyncableBaseEntityListFragment;
+import org.ovirt.mobile.movirt.ui.listfragment.SnapshotEmbeddableVmBoundResumeSyncableBaseEntityListFragment;
 import org.ovirt.mobile.movirt.util.CursorHelper;
 
 import java.util.List;
